+++ docs/guides/changelog.md
@@ -10,7 +10,7 @@ description: |-
 ## 2.0.0-beta2 (Unreleased)
 
 - (Enhancement) Add support for multiple `fmc_application_*` data sources and `fmc_application_filter` resource
-- (Enhancement) Add support for `fmc_application` and `fmc_application_filter` in Access Rules.
+- (Enhancement) Add support for Applications in Access Rules.
 
 ## 2.0.0-beta1